change temp_mean from Int to Float (2dp)

+++ api/app.py
@@ -39,7 +39,7 @@ class User(db.Model):
 class Reading(db.Model):
     reading_id = db.Column(db.Integer, primary_key=True)
     temp_mean = db.Column(db.Integer)
-    pressure_mean = db.Column(db.Integer)
+    pressure_mean = db.Column(db.Float)
     humidity_mean = db.Column(db.Integer)
     total_quality_mean = db.Column(db.Integer)
     sensor_id = db.Column(db.String(20), db.ForeignKey("user.sensor_id"))
